Split grouping tests by mechanism
Move the v1 grouping assertions into their own test module and extract
the v2 assertions from the broader issue tests. Keep transition test
names generic because the transition behavior is mechanism-agnostic.

See #440
